# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DCP Platform API
  version: 5.0.0
  description: |
    DCP is Saudi Arabia's first decentralized GPU compute marketplace.
    Providers register NVIDIA GPUs, install the Python daemon, and earn SAR.
    Renters submit compute jobs (LLM inference, image generation, training) that
    run on provider hardware. DCP takes 25% fee; providers earn 75%.

    ## Currency
    All monetary amounts are in **halala** (1 SAR = 100 halala) unless the field
    name ends in `_sar`.

    ## Auth
    Three auth schemes are used depending on the caller role:

    | Role | Scheme | How |
    |------|--------|-----|
    | Provider | `ProviderApiKey` | `?key=` query param or `x-provider-key` header |
    | Renter | `RenterApiKey` | `?key=` query param or `x-renter-key` header |
    | Admin | `AdminToken` | `x-admin-token` header |

    ## Rate Limits
    | Endpoint | Limit |
    |----------|-------|
    | `POST /providers/register` | 5 / IP / hour |
    | `POST /providers/heartbeat` | 4 / IP / minute |
    | `POST /renters/register` | 5 / IP / hour |
    | `POST /jobs/submit` | 10 / renter API key / minute |
    | `POST /payments/topup` | 10 / IP / minute |
    | `GET /api/providers/marketplace` | 30 / IP / minute |
    | `GET /api/admin/*` | 100 / IP / minute |
    | All other `/api/*` | 300 / IP / minute |

  contact:
    name: DCP DevRel
    url: https://dcp.sa/docs
  x-env-vars:
    DC1_ADMIN_TOKEN:
...
      required: true
    DC1_HMAC_SECRET:
...
      required: true
...
servers:
...
components:
  securitySchemes:
    ProviderApiKey:
      type: apiKey
      in: query
      name: key
      description: Provider API key. Prefix `dcp-provider-`. Also accepted as `x-provider-key` header.
    RenterApiKey:
      type: apiKey
      in: query
      name: key
      description: Renter API key. Prefix `dcp-renter-`. Also accepted as `x-renter-key` header.
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: Admin token set via DC1_ADMIN_TOKEN env var. Also accepted as Bearer token.
...
paths:
...
  /api/vllm/complete:
    post:
      tags: [vLLM]
      summary: Synchronous LLM completion
      description: |
        Submits a prompt to a managed vLLM inference endpoint and returns the
        full response synchronously. Internally submits an `llm_inference` job
        and waits for completion (max 120s timeout).

        For streaming output, use `POST /api/vllm/complete/stream`.
      security:
        - RenterApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, prompt]
              properties:
                model:
                  type: string
                  example: mistralai/Mistral-7B-Instruct-v0.2
                  description: Model ID from `/api/vllm/models`
                prompt:
                  type: string
                  example: Explain quantum computing in simple terms.
                max_tokens:
                  type: integer
                  default: 512
                  minimum: 1
                  maximum: 4096
                temperature:
                  type: number
                  default: 0.7
                  minimum: 0
                  maximum: 2
                top_p:
                  type: number
                  default: 1
                  minimum: 0
                  maximum: 1
                stop:
                  type: array
                  items:
                    type: string
                  description: Stop sequences
                provider_id:
                  type: integer
                  nullable: true
                  description: Pin to a specific provider. Omit to auto-select.
      responses:
        '200':
          description: Completion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Completion ID (same as job_id)
                  model:
                    type: string
                  choices:
                    type: array
                    items:
                      type: object
                      properties:
                        index:
                          type: integer
                        text:
                          type: string
                        finish_reason:
                          type: string
                          enum: [stop, length, timeout]
                  usage:
                    type: object
                    properties:
                      prompt_tokens:
                        type: integer
                      completion_tokens:
                        type: integer
                      total_tokens:
                        type: integer
                  cost_halala:
                    type: integer
                  latency_ms:
                    type: integer
        '402':
          description: Insufficient balance
        '503':
          description: No providers available for this model
  /api/vllm/complete/stream:
    post:
      tags: [vLLM]
      summary: Streaming LLM completion (SSE)
      description: |
        Submits a prompt and streams the response token-by-token as Server-Sent
        Events. Each event carries a partial `text` delta. The final event
        carries `finish_reason` and billing metadata.

        **Client usage:**
        ```javascript
        const resp = await fetch('https://api.dcp.sa/api/vllm/complete/stream', {
          method: 'POST',
          headers: {
            'Content-Type': 'application/json',
            'x-renter-key': apiKey
          },
          body: JSON.stringify({ model: 'mistralai/Mistral-7B-Instruct-v0.2', prompt: 'Hello!' })
        });
        const reader = resp.body.getReader();
        // read SSE frames...
        ```
      security:
        - RenterApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, prompt]
              properties:
                model:
                  type: string
                  example: mistralai/Mistral-7B-Instruct-v0.2
                prompt:
                  type: string
                max_tokens:
                  type: integer
                  default: 512
                temperature:
                  type: number
                  default: 0.7
                stop:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: SSE token stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Token delta events: `data: {"text": "...", "index": 0}\n\n`
                  Final event: `data: {"finish_reason": "stop", "usage": {...}, "cost_halala": N}\n\n`
                  Terminator: `data: [DONE]\n\n`
        '402':
          description: Insufficient balance
        '503':
          description: No providers available

  # ──────────────────────────────────────────────────────────────────────────────
  # Provider — GPU Profile, Withdrawals, Public Marketplace (DCP-322/332)
  # ──────────────────────────────────────────────────────────────────────────────
...